Split the kucoin types into separate files
