+++ packages/crypto-kraken-kucoin-api/src/index.ts
@@ -4,9 +4,9 @@ export {
     KuCoinTick,
     KuCoinListLanguages,
     KuCoinAllCoinsTick,
-    KuCoinOrderBook,
-    KuCoinBuyOrderBook,
-    KuCoinSellOrderBook,
+    KuCoinOrderBooks,
+    KuCoinBuyOrderBooks,
+    KuCoinSellOrderBooks,
     KuCoinResponseResult,
     KuCoinErrorResponseResult,
     KuCoinSuccessResponseResult,
