+++ packages/crypto-kraken-kucoin-api/src/kucoin-rest-v1/index.ts
@@ -1,3 +1,3 @@
 export * from './kucoin-constants';
 export * from './kucoin-rest-v1';
-export * from './kucoin-types';
+export * from './types';
