+++ packages/crypto-kraken-kucoin-api/src/kucoin-rest-v1/types/buy-order-books.ts
@@ -1,0 +1,15 @@
+import { FieldGuardsMap, isArray } from 'crypto-kraken-core';
+import { OrderBookOrder, orderBookOrderGuard } from './common';
+import { KuCoinSuccessResponseResult, kuCoinSuccessResponseResultGuardsMap } from './success-response-result';
+
+export interface KuCoinBuyOrderBooks extends KuCoinSuccessResponseResult {
+    data: OrderBookOrder[];
+}
+
+export const kuCoinBuyOrderBooksGuardsMap: FieldGuardsMap<KuCoinBuyOrderBooks> = {
+    ...kuCoinSuccessResponseResultGuardsMap,
+    data: {
+        this: isArray,
+        every: orderBookOrderGuard
+    }
+};
