+++ packages/crypto-kraken-kucoin-api/src/kucoin-rest-v1/types/error-response-result.ts
@@ -1,0 +1,15 @@
+import { FieldGuardsMap, isNumber, isString } from 'crypto-kraken-core';
+import { KuCoinResponseResult } from './response-result';
+
+export interface KuCoinErrorResponseResult extends KuCoinResponseResult {
+    success: false;
+    code: string;
+    msg: string;
+}
+
+export const kuCoinErrorResponseResultGuardsMap: FieldGuardsMap<KuCoinErrorResponseResult> = {
+    success: (value): value is KuCoinErrorResponseResult['success'] => value === false,
+    code: isString,
+    msg: isString,
+    timestamp: isNumber
+};
