+++ packages/crypto-kraken-kucoin-api/src/kucoin-rest-v1/types/order-type.ts
@@ -1,0 +1,9 @@
+export enum KuCoinOrderType {
+    SELL = 'SELL',
+    BUY = 'BUY'
+}
+
+export const isKuCoinOrderType = (data: any): data is KuCoinOrderType => {
+    /* istanbul ignore next */
+    return data === KuCoinOrderType.SELL || data === KuCoinOrderType.BUY;
+};
