+++ packages/crypto-kraken-kucoin-api/src/kucoin-rest-v1/types/sell-order-books.ts
@@ -1,0 +1,15 @@
+import { FieldGuardsMap, isArray } from 'crypto-kraken-core';
+import { OrderBookOrder, orderBookOrderGuard } from './common';
+import { KuCoinSuccessResponseResult, kuCoinSuccessResponseResultGuardsMap } from './success-response-result';
+
+export interface KuCoinSellOrderBooks extends KuCoinSuccessResponseResult {
+    data: OrderBookOrder[];
+}
+
+export const kuCoinSellOrderBooksGuardsMap: FieldGuardsMap<KuCoinSellOrderBooks> = {
+    ...kuCoinSuccessResponseResultGuardsMap,
+    data: {
+        this: isArray,
+        every: orderBookOrderGuard
+    }
+};
