+++ packages/crypto-kraken-kucoin-api/src/kucoin-rest-v1/types/success-response-result.ts
@@ -1,0 +1,15 @@
+import { FieldGuardsMap, isNumber, isString } from 'crypto-kraken-core';
+import { KuCoinResponseResult } from './response-result';
+
+export interface KuCoinSuccessResponseResult extends KuCoinResponseResult {
+    success: true;
+    code: 'OK';
+    msg: string;
+}
+
+export const kuCoinSuccessResponseResultGuardsMap: FieldGuardsMap<KuCoinSuccessResponseResult> = {
+    success: (value): value is KuCoinSuccessResponseResult['success'] => value === true,
+    code: (value): value is KuCoinSuccessResponseResult['code'] => value === 'OK',
+    msg: isString,
+    timestamp: isNumber
+};
